requirements.txt
# FastAPI framework
fastapi==0.115.0

# ASGI server to run FastAPI
uvicorn==0.30.6

# MongoDB async driver
motor==3.4.0
pymongo==4.8.0

# JWT authentication
pyjwt==2.9.0

# Password hashing (bcrypt)
passlib[bcrypt]==1.7.4

# Environment variable loader
python-dotenv==1.0.1

# Pydantic validation 
pydantic==2.8.2

# Pydantic with email support
pydantic[email]==2.8.2

pytest==8.3.3
pytest-asyncio==0.24.0

# Date and time utilities 
python-dateutil==2.9.0.post0

